+++ public/integrations.js
@@ -29,6 +29,10 @@
  * @property {0|1} facebookData.isActive
  * @property {string} facebookData.pixelId
  * @property {'both'|'pixel'|'capi'} facebookData.integrationType
+ * @property {Object} tiktokData
+ * @property {0|1} tiktokData.isActive
+ * @property {string} tiktokData.pixelId
+ * @property {'both'|'pixel'|'capi'} tiktokData.integrationType
  * @property {Object} gaData
  * @property {0|1} gaData.isActive
  * @property {string} gaData.gaCode
@@ -40,11 +44,15 @@
  */
 
 (() => {
+  let EXTERNAL_ID;
+
   const COOKIE = {
     UUID: "zc_uuid",
     COOKIE_CONSENT: 'cookieConsent',
     FBCLICK_ID: '_fbc',
     FBBROWSER_ID: '_fbp',
+    TTCLICK_ID: 'ttclid',
+    TTBROWSER_ID: '_ttp',
   }
 
   const consentCountries = [
@@ -54,6 +62,55 @@
   ];
 
   const logSyle = 'background: #2E495E;border-radius: 0.5em;color: white;font-weight: bold;padding: 2px 0.5em;';
+
+  getExternalId().then((id) => {EXTERNAL_ID = id});
+
+  function parseValue(value) {
+    if (value === 'true') return true;
+    else if (value === 'false') return false;
+    else if (value === undefined) return undefined;
+    else if (value === null) return null;
+    else if (isJSON(value)) return JSON.parse(value);
+    else if (!isNaN(Number(value))) return +value;
+    return value;
+  }
+
+  function isPlainObject(item) {
+    return (!!item && typeof item === 'object' && !Array.isArray(item));
+  }
+
+  function mergeDeep(target, ...sources) {
+    if (!sources.length) return { ...target };
+    const source = sources.shift();
+    const result = { ...target };
+
+    if (isPlainObject(result) && isPlainObject(source)) {
+      for (const key in source) {
+        if (isPlainObject(source[key])) {
+          result[key] = mergeDeep(result[key] || {}, source[key]);
+        } else {
+          result[key] = source[key];
+        }
+      }
+    }
+
+    return mergeDeep(result, ...sources);
+  }
+
+  async function sha256(message) {
+    // encode as (utf-8) Uint8Array
+    const msgBuffer = new TextEncoder().encode(message);                    
+
+    // hash the message
+    const hashBuffer = await crypto.subtle.digest('SHA-256', msgBuffer);
+
+    // convert buffer to byte array
+    const hashArray = Array.from(new Uint8Array(hashBuffer));
+
+    // convert bytes to hex string                  
+    const hashHex = hashArray.map(b => b.toString(16).padStart(2, '0')).join('');
+    return hashHex;
+  }
 
   /**
    * @param {string} str 
@@ -85,7 +142,7 @@
       if (cookie.indexOf(nameEQ) === 0) {
         const val = decodeURIComponent(cookie.substring(nameEQ.length));
 
-        if (!isNaN(parseFloat(val))) {
+        if (val.trim() !== '' && !isNaN(val) && !isNaN(parseFloat(val))) {
           return parseFloat(val);
         }
 
@@ -202,10 +259,43 @@
   }
 
   /**
+   * @param {Object} [params] - Optional parameters
+   * @param {string} [params.path] - The path
+   * @param {string} [params.key] - The cookie key
+   */
+  function getSession(params) {
+    const { path, key } = params || {};
+    const id = getCookie(key || COOKIE.UUID);
+
+    if (id) return { id };
+    
+    const sessionString = localStorage.getItem(key || COOKIE.UUID);
+    const sessions = (sessionString ? JSON.parse(atob(sessionString)) : null);
+    const pathName = path || globalThis?.location?.pathname || "/";
+    const session = sessions?.[pathName];
+    return session;
+  }
+
+  function getSessionId() {
+    const session = getSession({ key: COOKIE.UUID });
+    return session?.id || '';
+  }
+
+  function getEventId(eventName) {
+    return getSessionId() + '_' + eventName;
+  }
+
+  async function getExternalId() {
+    return sha256(getSessionId());
+  }  
+
+  /**
    * 
    * @param {string} cookieText 
+   * @param {string} countryCode
+   * @param {string} content_id
    */
-  function checkConsent(cookieText, countryCode) {
+  function checkConsent(cookieText, countryCode, content_id) {
     const cookieApp = window.VueCookieApp;
     if (!cookieApp || !cookieApp._) return false;
 
@@ -222,62 +312,336 @@
     cookiePopup.updateText(cookieText);
     
     // Show the cookie consent popup
-    cookiePopup.toggle(true);
+    cookiePopup.toggle(true, content_id);
   }
+
+  const Tiktok = {
+    options: {
+      id: null,
+      debug: false,
+      pageSlug: '',
+    },
+
+    // Temporary store for events triggered before cookie consent is given
+    temp: [],
+
+    log(...args) {
+      if (this.options.debug) {
+        console.log("%cTiktok Pixel", logSyle, ...args);
+      }
+    },
+
+    loadTemp() {
+      for (const args of this.temp) {
+        globalThis?.ttq?.[args[0]]?.(...args[1]);
+      }
+    },
+
+    getAMData() {
+      return {
+        external_id: EXTERNAL_ID,
+        country: Integration.data.countryCode,
+      }
+    },
+
+    /**
+     * 
+     * @param {string} eventName 
+     * @param {Record<string, any>} [params] 
+     * @param {Record<string, any>} [thirdArgs] 
+     * @param  {...any} restOfArgs 
+     * @returns 
+     */
+    event(eventName, params, thirdArgs, ...restOfArgs) {
+      const cookieConsent = parseValue(getCookie(COOKIE.COOKIE_CONSENT));
+      const paramObj = mergeDeep({...(params || {})}, this.getAMData());
+      const thirdArgsObj = mergeDeep({...(thirdArgs || {})}, { event_id: getEventId(eventName) });
+      const argList = [eventName, paramObj, thirdArgsObj, ...restOfArgs];
+
+      globalThis?.ttq?.track?.(...argList);
+
+      if (!cookieConsent && eventName !== 'Pageview') {
+        this.temp.push(['track', argList]);
+      }
+
+      this.log("event", argList);
+    },
+
+    track(...args) {
+      this.event(...args);
+    },
+
+    /**
+     * 
+     * @param {Object} payload 
+     * @param {string} [payload.email]
+     * @param {string} [payload.phone_number]
+     */
+    identify(payload) {
+      const cookieConsent = parseValue(getCookie(COOKIE.COOKIE_CONSENT));
+      const paramObj = mergeDeep({ ...(payload || {}) }, {
+        external_id: EXTERNAL_ID,
+      });
+
+      globalThis?.ttq?.identify?.(paramObj);
+
+      if (!cookieConsent) {
+        this.temp.push(['identify', [paramObj]]);
+      }
+
+      this.log("identify", paramObj);
+    },
+
+    /**
+     * @param {Object} payload
+     * @param {any} payload.value
+     * @param {any} payload.currency
+     * @param {string} payload.description
+     * @param {string} payload.content_id
+     * @param {string} payload.orderID
+     * @param {Object} payload.context
+     * @param {Object} [payload.contents]
+     * @param {string} payload.contents[].content_id
+     * @param {string} payload.contents[].content_name
+     * @param {number} payload.contents[].price
+     * @param {number} payload.contents[].quantity
+     */
+    purchase(payload) {
+      this.event('Purchase', {
+        quantity: 1,
+        content_type: 'product',
+        ...payload
+      });
+    },
+
+    /** 
+     * @param {string} siteUrl
+     * @param {Record<string, string>} [ttParams] - Optional TikTok parameters, if not provided, it will be retrieved from cookies and URL
+    */
+    prepareCAPIParams(siteUrl, ttParams) {
+      ttParams = ttParams || {
+        [COOKIE.TTCLICK_ID]: getCookie(COOKIE.TTCLICK_ID) || '',
+        [COOKIE.TTBROWSER_ID]: getCookie(COOKIE.TTBROWSER_ID) || ''
+      };
+
+      const location = new URL(`${siteUrl.startsWith('http') ? siteUrl : `https://${siteUrl}`}`);
+      const { ttclid } = parseQueryString(location?.search || '');
+
+      if (!ttParams[COOKIE.TTCLICK_ID] && ttclid) {
+        // const subdomainIndex = 1;
+        /* let subdomainIndex = location.host.split('.').length - 1;
+        if (subdomainIndex > 2) subdomainIndex = 2;
+        if (subdomainIndex < 0) subdomainIndex = 0; */
+    
+        ttParams[COOKIE.TTCLICK_ID] = ttclid || '';
+      }
+
+      return ttParams;
+    },
+
+    /**
+     * 
+     * @param {Object} params 
+     * @param {string} [params.siteUrl]
+     * @param {string} [params.value]
+     * @param {string} params.pageSlug
+     */
+    createCAPIClickID(params) {
+      const { value, pageSlug } = params;
+
+      return {
+        cookieName: COOKIE.TTCLICK_ID,
+        exdays: generateExpireTime(new Date(new Date().setDate(new Date().getDate() + 90))).date, // 90 days
+        value: value || '',
+        path: pageSlug
+      }
+    },
+
+    sendCAPIInfo() {
+      const win = globalThis;
+
+      setTimeout(() => {
+        const params = Tiktok.prepareCAPIParams(window.location.href);
+        const hasAnyValue = !!Object.values(params).filter(Boolean).length;
+
+        if (hasAnyValue) {
+          const myEvent = new CustomEvent('sendCAPIInfo', {
+            detail: {
+              integration: 'TT',
+              params: {
+                ttclid: params[COOKIE.TTCLICK_ID],
+                ttp: params[COOKIE.TTBROWSER_ID]
+              }
+            },
+            bubbles: true, 
+            cancelable: true
+          });
+
+          document.dispatchEvent(myEvent);
+        }
+      }, 1000);
+    },
+
+    /**
+     * @param {Object} payload
+     * @param {(number|string)} payload.pixelId
+     * @param {boolean} [payload.debug]
+     * @param {string} payload.pageSlug
+     * @returns 
+     */
+    init(payload) {
+      const headScripts = {
+        script: [],
+        noscript: []
+      };
+
+      if (this.options.id) return headScripts;
+      if (!payload.pixelId) {
+        console.warn("Tiktok Pixel cannot be installed because there is no Pixel ID!");
+        return headScripts;
+      }
+
+      this.options.id = payload.pixelId;
+      this.options.debug = !!payload.debug;
+      this.options.pageSlug = payload.pageSlug;
+
+      if (this.options.debug) {
+        this.track('load', this.options.id);
+        this.track('page');
+        return headScripts;
+      }
+
+      document.addEventListener('tiktokLoad', Tiktok.sendCAPIInfo.bind(this), { once: true });
+
+      if (window.document) {
+        (function (w, d, t, id, sid, ldtmp) {
+          w.TiktokAnalyticsObject=t;
+          var ttq=w[t]=w[t]||[];
+          ttq.methods=["page","track","identify","instances","debug","on","off","once","ready","alias","group","enableCookie","disableCookie"],
+          ttq.setAndDefer=function(t,e){t[e]=function(){t.push([e].concat(Array.prototype.slice.call(arguments,0)))}};
+          for(var i=0;i<ttq.methods.length;i++)ttq.setAndDefer(ttq,ttq.methods[i]);
+          ttq.instance=function(t){for(var e=ttq._i[t]||[],n=0;n<ttq.methods.length;n++)ttq.setAndDefer(e,ttq.methods[n]);return e},
+          ttq.load=function(e,n) {
+            var i="https://analytics.tiktok.com/i18n/pixel/events.js";ttq._i=ttq._i||{},ttq._i[e]=[],ttq._i[e]._u=i,ttq._t=ttq._t||{},
+            ttq._t[e]=+new Date,ttq._o=ttq._o||{},ttq._o[e]=n||{};var o=document.createElement("script");o.type="text/javascript",o.async=!0,
+            o.src=i+"?sdkid="+e+"&lib="+t;var a=document.getElementsByTagName("script")[0];a?.parentNode?.insertBefore(o,a);
+            document.dispatchEvent(new CustomEvent('tiktokLoad', {detail: {},bubbles: true, cancelable: true}));
+            w.loadedIntegrations=w.loadedIntegrations||[];
+            w.loadedIntegrations.push('Tiktok');
+          };
+          ttq.load(id, { historyObserver: false });
+          ttq.page(sid);
+          ttq.identify({ external_id: sid.external_id });
+          if (ldtmp) ldtmp();
+        }(window, document, 'ttq', this.options.id, {...this.getAMData(), event_id: getEventId('page') }, () => this.loadTemp()));
+      }
+
+      return headScripts
+    },
+  };
 
   const Facebook = {
     options: {
       id: null,
       debug: false,
       pageSlug: '',
-      countryCode: ''
     },
+
+    // Temporary store for events triggered before cookie consent is given
+    temp: [],
   
     log(...args) {
       if (this.options.debug) {
         console.log("%cFacebook Pixel", logSyle, ...args);
       }
     },
-    
-    event(...args) {
-      if (!this.options.id) return
-      if (window.fbq) {
-        window.fbq(...args);
+
+    loadTemp() {
+      for (const args of this.temp) {
+        globalThis?.fbq?.(...args);
       }
-      this.log("event", args);
     },
-  
+
+    getAMData() {
+      let clientIp;
+
+      try {
+        clientIp = Integration.data.ia ? atob(Integration.data.ia) : undefined
+      } catch {}
+
+      return {
+        external_id: EXTERNAL_ID,
+        country: Integration.data.countryCode,
+        client_ip_address: clientIp,
+        client_user_agent: globalThis?.navigator.userAgent,
+        fbp: getCookie(COOKIE.FBBROWSER_ID) || undefined,
+        fbc: getCookie(COOKIE.FBCLICK_ID) || undefined,
+      };
+    },
+    
+    /**
+     * @param {string} eventType
+     * @param {any} eventName
+     * @param {any} [params]
+     * @param {Record<string, any>} [forthArgs]
+     * @param {...any} restOfArgs
+     */
+    async event(eventType, eventName, params, forthArgs, ...restOfArgs) {
+      const cookieConsent = parseValue(getCookie(COOKIE.COOKIE_CONSENT));
+      const paramObj = mergeDeep((params || {}), this.getAMData());
+      const forthArgsObj = mergeDeep((forthArgs || {}), { eventID: getEventId(eventName) });
+      const argList = [eventType, eventName, paramObj, forthArgsObj, ...restOfArgs];
+
+      globalThis?.fbq?.(...argList);
+
+      if (!cookieConsent && eventName !== 'PageView' && eventType !== 'init') {
+        this.temp.push(argList);
+      }
+
+      this.log("event", argList);
+    },
+
+    /**
+     * @param {...any} args
+     */
     track(...args) {
-      if (!this.options.id) return
       this.event('track', ...args);
+    },
+
+    /**
+     * @param {...any} args
+     */
+    trackCustom(...args) {
+      this.event('trackCustom', ...args);
     },
     
     /**
      * @param {Object} payload
      * @param {string} payload.value
      * @param {string} payload.currency
-     * @param {string} payload.orderID
-     * @param {string} [payload.eventID]
+     * @param {Record<string, any>} payload.userData
+     * @param {Record<string, any>} [payload.otherData]
      */
     purchase(payload) {
       this.track('Purchase', {
         value: payload.value,
-        currency: payload.currency
-      }, {
-        eventID: payload.eventID,
-        orderID: payload.orderID,
-      })
+        currency: payload.currency,
+        pageUrl: window.location.href,
+        ...(payload.otherData || {}),
+        ...(payload.userData || {}),
+        fbp: getCookie(COOKIE.FBBROWSER_ID) || '',
+        fbc: getCookie(COOKIE.FBCLICK_ID) || ''
+      });
     },
   
-    prepareCAPIParams(siteUrl) {
-      const fbParams = {
+    prepareCAPIParams(siteUrl, fbParams) {
+      fbParams = fbParams || {
         [COOKIE.FBCLICK_ID]: getCookie(COOKIE.FBCLICK_ID) || '',
         [COOKIE.FBBROWSER_ID]: getCookie(COOKIE.FBBROWSER_ID) || ''
       };
   
       const location = new URL(`${siteUrl.startsWith('http') ? siteUrl : `https://${siteUrl}`}`);
-      const query = parseQueryString(location.search || '');
-      const fbclid = query.fbclid || ''
+      const { fbclid } = parseQueryString(location?.search || '');
   
       if (!fbParams[COOKIE.FBCLICK_ID] && fbclid) {
         const subdomainIndex = 1;
@@ -314,14 +678,21 @@
         const hasAnyValue = !!Object.values(params).filter(Boolean).length;
   
         if (hasAnyValue) {
-          sendFBCapi({
-            fbp: params[COOKIE.FBBROWSER_ID]
-          }, { hideErrorToast: true });
+          const myEvent = new CustomEvent('sendCAPIInfo', {
+            detail: {
+              integration: 'FB',
+              params: {
+                fbp: params[COOKIE.FBBROWSER_ID],
+                fbclid: params[COOKIE.FBCLICK_ID]
+              }
+            },
+            bubbles: true, 
+            cancelable: true
+          });
+
+          document.dispatchEvent(myEvent);
         }
       }, 1000);
-  
-      // Delete temp proxy object
-      if (win.proxy1) delete win.proxy1;
     },
   
     /**
@@ -329,7 +700,6 @@
      * @param {(number|string)} payload.pixelId
      * @param {boolean} [payload.debug]
      * @param {string} payload.pageSlug
-     * @param {string} payload.countryCode
      */
     init(payload) {
       const headScripts = {
@@ -346,13 +716,15 @@
       this.options.id = payload.pixelId;
       this.options.debug = !!payload.debug;
       this.options.pageSlug = payload.pageSlug;
-      this.options.countryCode = payload.countryCode;
   
       if (this.options.debug) {
         this.event('init', this.options.id);
         this.track('PageView');
         return headScripts;
       }
+
+      const win = globalThis;
+      document.addEventListener('fbLoad', Facebook.sendCAPIInfo.bind(this), { once: true });
   
       if (window.document) {
         (function(f, b, e, v) {
@@ -373,6 +745,7 @@
           t.src=v;
           t.onload=function() {
             const win = window;
+            document.dispatchEvent(new CustomEvent('fbLoad', {detail: {},bubbles: true,cancelable: true}));
             if (win) {
               win.loadedIntegrations=win.loadedIntegrations||[];
               win.loadedIntegrations.push('FB');
@@ -381,8 +754,9 @@
           const s = b.getElementsByTagName(e)[0];
           s.parentNode.insertBefore(t, s);
         })(window, document,'script', 'https://connect.facebook.net/en_US/fbevents.js');
-        this.event('init', this.options.id);
-        this.track('PageView');
+        this.event('init', this.options.id, this.getAMData());
+        this.track('PageView', this.getAMData());
+        this.loadTemp();
       }
   
       headScripts.noscript.push({
@@ -624,7 +998,8 @@
     /** @type {Integrations} */
     list: {},
     data: {
-      countryCode: ''
+      countryCode: '',
+      ia: ''
     },
     debug: false,
 
@@ -677,9 +1052,10 @@
       // Init if user access granted on cookie popup
       if (cookieConsent) {
         const metaScripts = this.meta();
-  
-        headScripts.script.push(...metaScripts.script);
-        headScripts.noscript.push(...metaScripts.noscript);
+        const tiktokScripts = this.tiktok();
+
+        headScripts.script.push(...metaScripts.script, ...tiktokScripts.script);
+        headScripts.noscript.push(...metaScripts.noscript, ...tiktokScripts.noscript);
       }
   
       return headScripts;
@@ -717,6 +1093,32 @@
       return headScripts;
     },
 
+    tiktok() {
+      const tiktokData = this.list.tiktokData;
+      const pageSlug = window.location.pathname || '';
+
+      const headScripts = {
+        script: [],
+        noscript: []
+      };
+
+      if (tiktokData?.isActive && tiktokData?.integrationType !== 'capi' && !!tiktokData?.pixelId) {
+        const payload = {
+          pixelId: tiktokData.pixelId,
+          debug: this.debug,
+          pageSlug,
+          countryCode: this.data.countryCode
+        };
+
+        const scripts = Tiktok.init(payload);
+
+        headScripts.script.push(...scripts.script);
+        headScripts.noscript.push(...scripts.noscript);
+      }
+
+      return headScripts;
+    },
+
     /**
      * 
      * @param {Array<{src?: string, async?: boolean; onload?: Function; innerHTML?: string}>} list 
@@ -738,11 +1140,12 @@
     }
   }
 
-  window.checkConsent = checkConsent;
-  window.setCookie = setCookie;
-  window.getCookie = getCookie;
-  window.Facebook = Facebook;
-  window.GTM = GTM;
-  window.GA4 = GA4;
-  window.Integration = Integration;
+  globalThis.checkConsent = checkConsent;
+  globalThis.setCookie = setCookie;
+  globalThis.getCookie = getCookie;
+  globalThis.Facebook = Facebook;
+  globalThis.GTM = GTM;
+  globalThis.GA4 = GA4;
+  globalThis.Tiktok = Tiktok;
+  globalThis.Integration = Integration;
 })();