+++ src/utils/handleUrlQuery.ts
@@ -1,4 +1,4 @@
-import { createPaymentSuccessForm } from "../lib/create";
+import { createPaymentSuccessForm, prepareButtonSuccessLink } from "../lib/create";
 import { type FormConfig, type IZotloCheckoutParams, type IZotloCardParams, PaymentCallbackStatus } from "../lib/types";
 import { handlePaymentSuccess } from "./sendPayment";
 
@@ -8,25 +8,48 @@ export enum UrlQuery {
   ERROR_MESSAGE = "zc_error_message",
 }
 
-export async function handleUrlQuery(payload: {
-  params: IZotloCheckoutParams | IZotloCardParams;
+export function getPaymentCallback(payload: {
   config: FormConfig;
 }) {
-  const { params, config } = payload || {};
   const queryString = globalThis?.location?.search || "";
   const urlParams = new URLSearchParams(queryString);
   const queryParams = Object.fromEntries(urlParams?.entries());
-  const status = config.cardUpdate
+  const status = payload.config.cardUpdate
     ? queryParams?.[UrlQuery.CARD_STATUS]
     : queryParams?.[UrlQuery.STATUS];
   const errorMessage = queryParams?.[UrlQuery.ERROR_MESSAGE] || "";
 
-  if (status === PaymentCallbackStatus.SUCCESS) {
+  return {
+    status,
+    errorMessage,
+    success: status === PaymentCallbackStatus.SUCCESS,
+    fail: status === PaymentCallbackStatus.FAIL,
+  };
+}
+
+export async function handleUrlQuery(payload: {
+  params: IZotloCheckoutParams | IZotloCardParams;
+  config: FormConfig;
+  reloadSession?: () => Promise<void>;
+}) {
+  const { params, config, reloadSession } = payload || {};
+  const { errorMessage, success, fail } = getPaymentCallback({ config });
+
+  if (success) {
     const paymentDetail = await handlePaymentSuccess({ config, params });
     if (paymentDetail) createPaymentSuccessForm({ config, paymentDetail });
+
+    if (!config.success.show) {
+      if (paymentDetail && config.general.isCheckoutLink) {
+        const redirectUrl = prepareButtonSuccessLink({ config, paymentDetail }) || '';
+        if (redirectUrl) window.location.href = redirectUrl;
+      }
+
+      await reloadSession?.();
+    }
   }
 
-  if (status === PaymentCallbackStatus.FAIL) {
+  if (fail) {
     params.events?.onFail?.({ message: errorMessage, data: {} });
   }
 }