+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 import Countries from '../countries.json';
-import { DesignTheme, type FormConfig, type IZotloCardParams, type IZotloCheckoutParams, PaymentProvider, PaymentResultStatus, SavedCardsGroupName, type FooterInfo, type FormSetting, PackageType } from '../lib/types';
+import { DesignTheme, type FormConfig, type IZotloCardParams, type IZotloCheckoutParams, PaymentProvider, PaymentResultStatus, SavedCardsGroupName, type FooterInfo, type FormSetting, PackageType, PackageCondition } from '../lib/types';
 import { createAllCardsModal, createPriceTable, createSavedCardItem, createButton } from '../lib/create';
 import { getPackageTemplateParams } from './getPackageInfo';
 import { getCDNUrl } from './getCDNUrl';
@@ -337,10 +337,26 @@ export function handleSavedCardsEvents(params: { config: FormConfig }) {
 
 export function getFooterPriceInfo(config: FormConfig) {
   const { $t } = useI18n(config?.general?.localization);
+  const params = getPackageTemplateParams(config);
 
-  if (config.paymentData?.package?.packageType === PackageType.SUBSCRIPTION) return '';
-  
-  return template($t('footer.legals.oneTimeInfo'), getPackageTemplateParams(config));
+  if (config.general.isActiveFTC) {
+    if (config.paymentData?.package?.packageType === PackageType.SUBSCRIPTION) return '';
+    return template($t('footer.legals.oneTimeInfo'), params);
+  }
+
+  const packageCondition = config?.packageInfo?.condition || PackageCondition.PACKAGE_WITH_TRIAL;
+  const isDiscountCodeApplied = getIsDiscountCodeApplied(config);
+  const isTrialDiscountAllowed = !!config?.paymentData?.discount?.allowTrial || false;
+  const isRecurringDiscountLimited = (config?.paymentData?.discount?.recurringMode === 'limited') || false;
+
+  const discountTrialKey = isTrialDiscountAllowed ? 'allowTrialDiscount' : 'noTrialDiscount';
+  const discountRecurringKey = isRecurringDiscountLimited ? 'recurringLimited' : 'recurringForever';
+  const finalLocalizationKey = isDiscountCodeApplied ? 
+    `footer.priceInfo.discounted.${packageCondition}.${discountTrialKey}.${discountRecurringKey}` : 
+    `footer.priceInfo.${packageCondition}`;
+
+  return template($t(finalLocalizationKey), params);
+
 }
 
 export function getSubmitButtonContent(config: FormConfig) {
@@ -386,14 +402,16 @@ export async function handlePriceChanges(config: FormConfig) {
     });
   }
 
-  updateElementsValue<HTMLElement>('[data-total-price]', config?.packageInfo?.totalPayableAmount);
+  updateElementsValue<HTMLElement>('[data-total-price]', config?.packageInfo?.totalPayableAmount as string);
   updateElementsValue<HTMLButtonElement>('[data-card-submit-button]', getSubmitButtonContent(config));
   updateElementsValue<HTMLElement>('[data-original-price]', config?.packageInfo?.discount?.original as string);
-  updateElementsValue<HTMLElement>('[data-discount-price]', config?.packageInfo?.discount?.price as string);
   const footerFullDescription = getFooterPriceInfo(config);
-  const priceTable = createPriceTable({config});
   updateElementsValue<HTMLElement>('[data-footer-description]', footerFullDescription);
-  updateElementsValue<HTMLElement>('[data-price-table]', priceTable);
+
+  if (config.general.canViewPriceTable) {
+    const priceTable = createPriceTable({config});
+    updateElementsValue<HTMLElement>('[data-price-table]', priceTable);
+  }
 }
 
 export function getIsDiscountCodeApplied(config: FormConfig): boolean {
@@ -525,5 +543,45 @@ export async function handleResponseRedirection(payload: {
       }
       globalThis.location.href = returnUrl;
     }
+  }
+}
+
+async function sha256(message: string) {
+  // encode as (utf-8) Uint8Array
+  const msgBuffer = new TextEncoder().encode(message);                    
+
+  // hash the message
+  const hashBuffer = await crypto.subtle.digest('SHA-256', msgBuffer);
+
+  // convert buffer to byte array
+  const hashArray = Array.from(new Uint8Array(hashBuffer));
+
+  // convert bytes to hex string                  
+  const hashHex = hashArray.map(b => b.toString(16).padStart(2, '0')).join('');
+  return hashHex;
+}
+
+export async function getUserDataForIntegration(payload: {
+  registerType: string;
+  subscriberId?: string;
+}) {
+  const { registerType, subscriberId } = payload || {};
+  const hashedSubscriberId = subscriberId ? await sha256(subscriberId) : undefined;
+  const isEmail = subscriberId && (''+subscriberId.includes('@') || ''+subscriberId.includes('@privaterelay.appleid.com'));
+  const registerFinalType = registerType === 'other'
+    ? isEmail ? 'email' : undefined
+    : registerType;
+
+  const userData: Record<string, any> = {
+    ...(registerFinalType && hashedSubscriberId ? {[registerFinalType === 'email' ? 'em' : 'ph']: hashedSubscriberId} : {}),
+  }
+
+  const context = {
+    ...(registerFinalType && hashedSubscriberId ? {[registerFinalType === 'email' ? 'email' : 'phone_number']: hashedSubscriberId} : {})
+  }
+
+  return {
+    user_data: userData,
+    context
   }
 }